requirements.txt
# --- AI & Computer Vision ---
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2
tensorflow-macos==2.15.0
tensorflow-metal==1.2.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0

# --- Web Backend (FastAPI) ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
aiofiles==23.2.1
loguru==0.7.2

# --- Data Science & Utilities ---
pandas==2.1.4
scikit-learn==1.4.0
scikit-image==0.22.0
albumentations==1.3.1
imgaug==0.4.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
tqdm==4.66.1
pyyaml==6.0.1
requests==2.31.0

# --- Jupyter & Notebooks ---
jupyter==1.0.0
ipykernel==6.29.0
notebook==7.0.6

# --- Database Drivers (Optional) ---
pymysql==1.1.0
cryptography==41.0.7